src/server/requirements.txt
Flask
flask_cors
gunicorn
requests
mysql-connector-python
python-dotenv
